Introduce data_directory and output_directory
NOTE: Unlike the other PRs I recently sent out, this one is actually a
functional change. We allow the user to select a location to read data
from as well as to write output images to. The output directory is by
default /tmp/data.

Internally, the script lacks permissions to write to the machine that is
running the script. Hence, we need to add these flags.